added real data to forecast screen, fixed timezone forecast issues

+++ data/src/main/java/com/lab/esh1n/data/cache/entity/WeatherWithPlace.kt
@@ -27,7 +27,7 @@ class WeatherWithPlace(
 
         var description: String,
 
-        var windSpeed: Float,
+        var windSpeed: Double,
 
         var windDegree: Int,
